LayoutTests/fast/forms/state-restore-various-values.html

<!DOCTYPE html>
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p>Test if special characters are correctly restored.</p>
<div id="console"></div>

<input id="emptyOnFirstVisit">
<div id="parent">
<form action="data:text/html,&lt;script>history.back()&lt;/script>" id=form1>
<select id="select1" multiple>
  <option id="opt-01"></option>
  <option id="opt-02">,</option>
  <option id="opt-03">\</option>
  <option id="opt-04">,a,</option>
  <option id="opt-05">,\,</option>
  <option id="opt-06">,\\,</option>
  <option id="opt-07">\a\</option>
  <option id="opt-08">\n\</option>
  <option id="opt-09">\,,\</option>
  <option id="opt-10">&amp;&#0;&#x0d;&#x0a;,</option>
</select>
</form>
</div>

<script>
// Note that this test depends on the fact that select options are stored by
// value strings, not indexes.

jsTestIsAsync = true;

function runTest()
{
    var state = document.getElementById('emptyOnFirstVisit');
    if (!state.value) {
        // First visit.
        setTimeout(function() {
            state.value = 'visited';
            var options = $('select1').options;
            for (var i = 0; i < options.length; ++i)
                options[i].selected = true;
            $('form1').submit();
        }, 0);
    } else {
        // Went back to this page again, and form state should be restored.
        shouldBeTrue('$("opt-01").selected');
        shouldBeTrue('$("opt-02").selected');
        shouldBeTrue('$("opt-03").selected');
        shouldBeTrue('$("opt-04").selected');
        shouldBeTrue('$("opt-05").selected');
        shouldBeTrue('$("opt-06").selected');
        shouldBeTrue('$("opt-07").selected');
        shouldBeTrue('$("opt-08").selected');
        shouldBeTrue('$("opt-09").selected');
        shouldBeTrue('$("opt-10").selected');
    
        $('parent').innerHTML = '';
        setTimeout(function() {
            finishJSTest();
        }, 0);
    }
}

runTest();
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
